Mainly fixed with some problems still

+++ app/Controllers/TripsController.js
@@ -1,12 +1,13 @@
 import { ProxyState } from "../AppState.js";
 import { tripsService } from "../Services/TripsService.js";
 import { loadState, saveState } from "../Utils/LocalStorage.js";
+import { Pop } from "../Utils/Pop.js";
 
 
 function _drawTrips() {
     let trips = ProxyState.trips
     let template = ''
-    trips.forEach(t => { template += t.Template })
+    trips.forEach(t => template += t.Template)
     document.getElementById('trips').innerHTML = template
 }
 
@@ -21,13 +22,27 @@ export class TripsController {
         _drawTrips()
     }
 
+    // TODO add create trip function
+    createTrip() {
+        window.event.preventDefault()
+        console.log('creating a trip');
+        let form = window.event.target
+        let tripData = {
+            name: form.name.value,
+        }
+        console.log('tripData', tripData);
+        tripsService.createTrip(tripData)
+    }
+
     updateTrip(id) {
         let textarea = window.event.target
         console.log(textarea.value, id);
         tripsService.updateTrip(textarea.value, id)
     }
 
-    deleteTrip(id) {
-        tripsService.deleteTrip(id)
+    async deleteTrip(id) {
+        if (await Pop.confirm('are you sure you want to delete this?')) {
+            tripsService.deleteTrip(id)
+        }
     }
 }
